Finish initial layout code
