+++ forui/lib/foundation.dart
@@ -3,4 +3,4 @@ library forui.foundation;
 
 export 'src/foundation/layout.dart';
 export 'src/foundation/portal/portal.dart';
-export 'src/foundation/portal/portal_shift.dart' hide Alignments;
+export 'src/foundation/portal/portal_shift.dart';
